pyproject.toml
[project]
name = "panel-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-admin>=2.0.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "waitress>=3.0.2",
    "wtforms>=3.2.1",
]

[tool.uv]
python-install-mirror = "https://github-proxy.lixxing.top/https://github.com/astral-sh/python-build-standalone/releases/download"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
